[TP1-002/3/4] - { feat }: se realizan cambios

+++ src/ejercicios/ejercicio1103.py
@@ -11,13 +11,13 @@ y para mostrar el resultado con decimales en el ultimo punto lo lograremos de la
 "print("El ", numero2, "% " "de ",numero1, "es", numero1 * (numero2/100))"
 """
 division = 0.0
-suma = 0.0
-porcentaje = 0.0
+suma = 0
+porcentaje = 0
 divisionPosible = False
 numero1 = int(input("Ingrese primer numero entero:"))
 numero2 = int(input("Ingrese segundo numero entero:"))
 suma = numero1 + numero2
-porcentaje = numero1 * numero2/100 
+porcentaje = int(numero1 * numero2/100)
 if numero2 != 0:
     division = numero1/numero2   
     divisionPosible = numero1%numero2==0
